backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
Pillow==10.4.0
cairosvg==2.7.1
icnsutil==1.1.0
